LIVY-306. Removed hard-coded minimum timeout value. (#290)
Currently a session timeout can't be configured below 1h. This is due to a hard-coded value that was missed in the clean up of LIVY-114 and LIVY-127.

Removed the limit and the conf value (default of 1h) will be used.
(cherry picked from commit a72bbff6fbf961f7413f3ca87d2a244f182f98ff)

Signed-off-by: jerryshao <sshao@hortonworks.com>